fix: process the aggregated response in aggregator (#705)
