--- types/generics.go
@@ -1,31 +1,0 @@
-package types
-
-import (
-	"math/big"
-)
-
-type TaskResponseMetadata struct {
-	TaskRespondedBlock uint32
-	HashOfNonSigners   [32]byte
-}
-
-type BN254G1Point struct {
-	X *big.Int
-	Y *big.Int
-}
-
-type BN254G2Point struct {
-	X [2]*big.Int
-	Y [2]*big.Int
-}
-
-type NonSignerStakesAndSignature struct {
-	NonSignerQuorumBitmapIndices []uint32
-	NonSignerPubkeys             []BN254G1Point
-	QuorumApks                   []BN254G1Point
-	ApkG2                        BN254G2Point
-	Sigma                        BN254G1Point
-	QuorumApkIndices             []uint32
-	TotalStakeIndices            []uint32
-	NonSignerStakeIndices        [][]uint32
-}
