+++ types/task_manager.go
@@ -1,0 +1,52 @@
+package types
+
+import (
+	"math/big"
+
+	"github.com/Layr-Labs/eigensdk-go/crypto/bls"
+)
+
+type TaskResponseMetadata struct {
+	TaskRespondedBlock uint32
+	HashOfNonSigners   [32]byte
+}
+
+type NonSignerStakesAndSignature struct {
+	NonSignerQuorumBitmapIndices []uint32
+	NonSignerPubkeys             []BN254G1Point
+	QuorumApks                   []BN254G1Point
+	ApkG2                        BN254G2Point
+	Sigma                        BN254G1Point
+	QuorumApkIndices             []uint32
+	TotalStakeIndices            []uint32
+	NonSignerStakeIndices        [][]uint32
+}
+
+type BN254G1Point struct {
+	X *big.Int
+	Y *big.Int
+}
+
+type BN254G2Point struct {
+	X [2]*big.Int
+	Y [2]*big.Int
+}
+
+// These conversion utils are similar to the ones that can be found on chainio/utils/utils.go. The only difference
+// is that the BN254 point type in that case is from the registry coordinator binding and in this case is the type
+// We use to communicate with the task manager contract
+func ConvertToBN254G1Point(input *bls.G1Point) BN254G1Point {
+	output := BN254G1Point{
+		X: input.X.BigInt(big.NewInt(0)),
+		Y: input.Y.BigInt(big.NewInt(0)),
+	}
+	return output
+}
+
+func ConvertToBN254G2Point(input *bls.G2Point) BN254G2Point {
+	output := BN254G2Point{
+		X: [2]*big.Int{input.X.A1.BigInt(big.NewInt(0)), input.X.A0.BigInt(big.NewInt(0))},
+		Y: [2]*big.Int{input.Y.A1.BigInt(big.NewInt(0)), input.Y.A0.BigInt(big.NewInt(0))},
+	}
+	return output
+}
